new try

+++ hashmap.c
@@ -40,7 +40,7 @@ int is_equal(void* key1, void* key2){
 
 
 void insertMap(HashMap * map, char * key, void * value) {
-    int i = hash(key,map->capacity);
+    long i = hash(key,map->capacity);
 
     while(map->buckets[i]->key != NULL && map->buckets[i] != NULL)
     {
